requirements.txt
aws-cdk-lib==2.113.0
constructs>=10.0.0,<11.0.0
aws-cdk-constructs==2.22.2
